+++ packages/client/src/observable/internal/base-list/BaseListQuery.ts
@@ -598,7 +598,7 @@ export abstract class BaseListQuery<
     this.store.batch({}, (batch) => {
       const objectCacheKey = this.store.cacheKeys.get(
         "object",
-        object.$apiName,
+        object.$objectType ?? object.$apiName,
         object.$primaryKey,
       );
       batch.delete(objectCacheKey, "loaded");