+++ packages/generator/src/v2.0/generateClientSdkVersionTwoPointZero.test.ts
@@ -592,7 +592,7 @@ describe("generator", () => {
             /**
              * Todo(s) to be deleted
              */
-            readonly object?: ReadonlyArray<ActionParam.ObjectType<Todo>>;
+            readonly object?: ReadonlyArray<ActionParam.ObjectType<Todo>> | null;
           }
 
           // Represents a fqn of the action
@@ -614,6 +614,11 @@ describe("generator", () => {
 
         /**
          * An action which takes in an array of objects
+         *
+         * **Note on null values:** _For optional parameters, explicitly providing a null value instead of undefined
+         * can change the behavior of the applied action. If prefills are configured, null prevents them
+         * from being applied. If a parameter modifies an object's property, null will clear the data from
+         * the object, whereas undefined would not modify that property._
          * @param {ActionParam.ObjectType<Todo>} [object] Todo(s) to be deleted
          */
         export interface deleteTodos extends ActionDefinition<deleteTodos.Signatures> {
@@ -668,7 +673,7 @@ describe("generator", () => {
             /**
              * A Todo to mark completed
              */
-            readonly object?: ActionParam.ObjectType<Todo>;
+            readonly object?: ActionParam.ObjectType<Todo> | null;
           }
 
           // Represents a fqn of the action
@@ -690,6 +695,11 @@ describe("generator", () => {
 
         /**
          * An action which takes different types of parameters
+         *
+         * **Note on null values:** _For optional parameters, explicitly providing a null value instead of undefined
+         * can change the behavior of the applied action. If prefills are configured, null prevents them
+         * from being applied. If a parameter modifies an object's property, null will clear the data from
+         * the object, whereas undefined would not modify that property._
          * @param {ActionParam.ObjectType<Todo>} [object] A Todo to mark completed
          */
         export interface markTodoCompleted extends ActionDefinition<markTodoCompleted.Signatures> {
@@ -790,11 +800,14 @@ describe("generator", () => {
           };
         }
 
-        export const SomeInterface: SomeInterface = {
+        export const SomeInterface = {
           type: 'interface',
           apiName: 'SomeInterface',
           osdkMetadata: $osdkMetadata,
-        };
+          internalDoNotUseMetadata: {
+            rid: 'idk',
+          },
+        } satisfies SomeInterface & { internalDoNotUseMetadata: { rid: string } } as SomeInterface;
         ",
           "/foo/ontology/objects.ts": "export { Person } from './objects/Person.js';
         export { Todo } from './objects/Todo.js';
@@ -1255,7 +1268,7 @@ describe("generator", () => {
             /**
              * Todo(s) to be deleted
              */
-            readonly object?: ReadonlyArray<ActionParam.ObjectType<Todo>>;
+            readonly object?: ReadonlyArray<ActionParam.ObjectType<Todo>> | null;
           }
 
           // Represents a fqn of the action
@@ -1277,6 +1290,11 @@ describe("generator", () => {
 
         /**
          * An action which takes in an array of objects
+         *
+         * **Note on null values:** _For optional parameters, explicitly providing a null value instead of undefined
+         * can change the behavior of the applied action. If prefills are configured, null prevents them
+         * from being applied. If a parameter modifies an object's property, null will clear the data from
+         * the object, whereas undefined would not modify that property._
          * @param {ActionParam.ObjectType<Todo>} [object] Todo(s) to be deleted
          */
         export interface deleteTodos extends ActionDefinition<deleteTodos.Signatures> {
@@ -1331,7 +1349,7 @@ describe("generator", () => {
             /**
              * A Todo to mark completed
              */
-            readonly object?: ActionParam.ObjectType<Todo>;
+            readonly object?: ActionParam.ObjectType<Todo> | null;
           }
 
           // Represents a fqn of the action
@@ -1353,6 +1371,11 @@ describe("generator", () => {
 
         /**
          * An action which takes different types of parameters
+         *
+         * **Note on null values:** _For optional parameters, explicitly providing a null value instead of undefined
+         * can change the behavior of the applied action. If prefills are configured, null prevents them
+         * from being applied. If a parameter modifies an object's property, null will clear the data from
+         * the object, whereas undefined would not modify that property._
          * @param {ActionParam.ObjectType<Todo>} [object] A Todo to mark completed
          */
         export interface markTodoCompleted extends ActionDefinition<markTodoCompleted.Signatures> {
@@ -1453,11 +1476,14 @@ describe("generator", () => {
           };
         }
 
-        export const SomeInterface: SomeInterface = {
+        export const SomeInterface = {
           type: 'interface',
           apiName: 'foo.bar.SomeInterface',
           osdkMetadata: $osdkMetadata,
-        };
+          internalDoNotUseMetadata: {
+            rid: 'idk',
+          },
+        } satisfies SomeInterface & { internalDoNotUseMetadata: { rid: string } } as SomeInterface;
         ",
           "/foo/ontology/objects.ts": "export { Person } from './objects/Person.js';
         export { Todo } from './objects/Todo.js';
@@ -2719,11 +2745,14 @@ describe("generator", () => {
           };
         }
 
-        export const SomeInterface: SomeInterface = {
+        export const SomeInterface = {
           type: 'interface',
           apiName: 'com.example.dep.SomeInterface',
           osdkMetadata: $osdkMetadata,
-        };
+          internalDoNotUseMetadata: {
+            rid: 'idk2',
+          },
+        } satisfies SomeInterface & { internalDoNotUseMetadata: { rid: string } } as SomeInterface;
         ",
           "/foo/ontology/objects.ts": "export { Task } from './objects/Task.js';
         ",