+++ packages/react-components/src/object-table/TableBody.tsx
@@ -17,6 +17,7 @@
 import type { Cell, Row, RowData } from "@tanstack/react-table";
 import { useVirtualizer } from "@tanstack/react-virtual";
 import React, { useLayoutEffect } from "react";
+import styles from "./TableBody.module.css";
 import { TableRow } from "./TableRow.js";
 
 interface TableBodyProps<TData extends RowData> {
@@ -52,9 +53,8 @@ export function TableBody<TData extends RowData>({
 
   return (
     <tbody
+      className={styles.osdkTableBody}
       style={{
-        display: "grid",
-        position: "relative",
         height: `${rowVirtualizer.getTotalSize()}px`,
       }}
     >