+++ packages/widget.client-react/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @osdk/widget.client-react
 
+## 3.4.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies [3fc5fe6]
+- Updated dependencies [bb9d25c]
+  - @osdk/client@2.7.0-beta.12
+  - @osdk/widget.client@3.4.0-beta.12
+
+## 3.4.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [6cfe14a]
+- Updated dependencies [ec6ad57]
+- Updated dependencies [38d5958]
+- Updated dependencies [d5cfc38]
+  - @osdk/client@2.7.0-beta.11
+  - @osdk/widget.client@3.4.0-beta.11
+
+## 3.4.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies [db44f6b]
+- Updated dependencies [266b901]
+  - @osdk/client@2.7.0-beta.10
+  - @osdk/widget.client@3.4.0-beta.10
+
 ## 3.4.0-beta.9
 
 ### Patch Changes